🆕 chore(.env.example): add example environment variables This commit adds an example .env file with environment variables that need to be set up for the application to work correctly. The variables include GOOGLE_ID, GOOGLE_CLIENT_SECRET, MONGODB_URI, NEXTAUTH_URL, NEXTAUTH_URL_INTERNAL, and NEXTAUTH_SECRET. These variables are necessary for the application to connect to the Google API, MongoDB, and NextAuth.
🎉 feat(README.md): Add documentation for the project
This commit adds documentation for the project, including the project's purpose, requirements, technologies used, file structure, and more. It also explains why Next.js was chosen over pure React and how the routing system works in Next.js. Additionally, it provides information on the API folder, component folder, and model folder, including the Pun model and the User model. Finally, it includes an .env.example file and information on how to configure the Callback URI for the OAuth client on console.google.com.

🎉 feat(ReadmeFiles): add file structure and uri config images
Added two new images to the ReadmeFiles directory. The fileStructure.png image shows the directory structure of the project, while the uriConfig.png image shows the configuration of the URI used in the project.